+++ packages/@headlessui-react/src/components/mouse.ts
@@ -1,0 +1,4 @@
+export enum MouseButton {
+  Left = 0,
+  Right = 2,
+}